Initial commit of entire project
All code for the OpenFlow implementation. The Terminal class, the constructor of the Node class and the Listener class were implemented to assist with user I/O and threading. The remainder of the Node class as well as the Controller, Switch and EndNode classes were written by me to model a Software Defined Network using OpenFlow.